backend/requirements.txt
alembic==1.17.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==1.26.4
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
pmdarima==2.0.4
psycopg2-binary==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
SQLAlchemy==2.0.44
statsmodels==0.14.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3